Refactor things into their own files
