+++ src/traits/mod.rs
@@ -2,8 +2,7 @@
 // SPDX-License-Identifier: BlueOak-1.0.0 OR BSD-2-Clause-Patent
 // SPDX-FileContributor: Piper McCorkle <piper@lutris.engineering>
 
-mod attribute_iter;
-pub use attribute_iter::*;
+// For clarity's sake, this is in alphabetical order.
 
 mod connection;
 pub use connection::*;
@@ -11,15 +10,5 @@ pub use connection::*;
 mod database;
 pub use database::*;
 
-mod datom_iter;
-pub use datom_iter::*;
-
 mod entity;
 pub use entity::*;
-
-/// C bindings
-#[cfg(feature = "c")]
-pub mod c {
-    pub use super::connection::c::*;
-    pub use super::database::c::*;
-}
